cleaned up unit tests and consolidated code
