+++ nonhl7-orchestration/fns/fns-csv-pipeline/src/test/kotlin/gov/cdc/dex/csv/functions/activity/Unit_FnCSVValidationGeneric.kt
@@ -9,6 +9,7 @@ import gov.cdc.dex.csv.dtos.ActivityInput
 import gov.cdc.dex.csv.dtos.ActivityParams
 import gov.cdc.dex.csv.dtos.CommonInput
 import gov.cdc.dex.csv.services.IBlobService
+import gov.cdc.dex.csv.BlobServiceMocker
 
 import java.io.File
 import java.io.InputStream
@@ -29,7 +30,7 @@ import org.mockito.invocation.InvocationOnMock
 
 internal class Unit_FnCSVValidationGeneric {
     companion object{
-        private val outputParentDir = File("target/test-output/"+LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMddHHmmssSSSSS")))
+        private val outputParentDir = File("target/test-output/Unit_FnCSVValidationGeneric/"+LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMddHHmmssSSSSS")))
     }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -38,7 +39,7 @@ internal class Unit_FnCSVValidationGeneric {
     internal fun happyPath_singleCsv(){
         moveToOutput("test-upload.csv", "happyPath_singleCsv")
         val input = ActivityInput(common=CommonInput("happyPath_singleCsv", ActivityParams(originalFileUrl = "happyPath_singleCsv/test-upload.csv")))
-        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), mockBlobService())
+        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), BlobServiceMocker.mockBlobService(outputParentDir))
         Assertions.assertNull(response.updatedParams)
         Assertions.assertNull(response.errorMessage)
         Assertions.assertNull(response.fanOutParams)
@@ -50,7 +51,7 @@ internal class Unit_FnCSVValidationGeneric {
     @Test
     internal fun negative_url_null(){
         val input = ActivityInput(common=CommonInput("negative_url_null", ActivityParams()))
-        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), mockBlobService())
+        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), BlobServiceMocker.mockBlobService(outputParentDir))
         Assertions.assertEquals("No source URL provided!", response.errorMessage)
         Assertions.assertNull(response.updatedParams)
         Assertions.assertNull(response.fanOutParams)
@@ -59,7 +60,7 @@ internal class Unit_FnCSVValidationGeneric {
     @Test
     internal fun negative_url_empty(){
         val input = ActivityInput(common=CommonInput("negative_url_empty", ActivityParams(originalFileUrl = "")))
-        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), mockBlobService())
+        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(),BlobServiceMocker.mockBlobService(outputParentDir))
         Assertions.assertEquals("No source URL provided!", response.errorMessage)
         Assertions.assertNull(response.updatedParams)
         Assertions.assertNull(response.fanOutParams)
@@ -68,7 +69,7 @@ internal class Unit_FnCSVValidationGeneric {
     @Test
     internal fun negative_url_notFound(){
         val input = ActivityInput(common=CommonInput("negative_url_notFound", ActivityParams(originalFileUrl = "some-other-file.csv")))
-        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), mockBlobService())
+        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), BlobServiceMocker.mockBlobService(outputParentDir))
         Assertions.assertEquals("File does not exist!", response.errorMessage)
         Assertions.assertNull(response.updatedParams)
         Assertions.assertNull(response.fanOutParams)
@@ -81,7 +82,7 @@ internal class Unit_FnCSVValidationGeneric {
     internal fun negative_file_notCSV(){
         moveToOutput("test-upload.xls","negative_file_notCSV")
         val input = ActivityInput(common=CommonInput("negative_file_unableToZip", ActivityParams(originalFileUrl = "negative_file_unableToZip/test-upload.xls")))
-        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), mockBlobService())
+        val response = FnCSVValidationGeneric().process(input, ContextMocker.mockExecutionContext(), BlobServiceMocker.mockBlobService(outputParentDir))
         Assertions.assertEquals("File is not a .csv!", response.errorMessage)
         Assertions.assertNull(response.updatedParams)
         Assertions.assertNull(response.fanOutParams)
@@ -96,41 +97,5 @@ internal class Unit_FnCSVValidationGeneric {
         
         localFileOut.parentFile.mkdirs()
         localFileIn.copyTo(localFileOut)
-    }
-
-    private fun mockBlobService(contentType:String = "DUMMY_CONTENT"):IBlobService{
-        val mockBlobService = Mockito.mock(IBlobService::class.java)
-
-        Mockito.`when`(mockBlobService.exists(Mockito.anyString())).thenAnswer(this::doesTestFileExist)
-        Mockito.`when`(mockBlobService.getProperties(Mockito.anyString())).thenAnswer({getBlobProperties(it, contentType)})
-        Mockito.`when`(mockBlobService.openDownloadStream(Mockito.anyString())).thenAnswer(this::openInputStream)
-        Mockito.`when`(mockBlobService.openUploadStream(Mockito.anyString())).thenAnswer(this::openOutputStream)
-
-        return mockBlobService
-    }
-
-    private fun doesTestFileExist(i: InvocationOnMock):Boolean{
-        val relativePath:String = i.getArgument(0)
-        var localFile = File(outputParentDir,relativePath)
-        return localFile.exists()
-    }
-
-    private fun getBlobProperties(i: InvocationOnMock, contentType:String):BlobProperties{
-        val relativePath:String = i.getArgument(0)
-
-        return BlobProperties(null, null, null, -1, contentType, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null, null)
-    }
-
-    private fun openInputStream(i: InvocationOnMock):InputStream{
-        val relativePath:String = i.getArgument(0)
-        var localFile = File(outputParentDir,relativePath)
-        return FileInputStream(localFile)
-    }
-
-    private fun openOutputStream(i: InvocationOnMock):OutputStream{
-        val relativePath:String = i.getArgument(0)
-        var localFile = File(outputParentDir,relativePath)
-        localFile.parentFile.mkdirs()
-        return FileOutputStream(localFile)
     }
 }
